+++ artist/scenario/surface_generator.py
@@ -2,10 +2,12 @@ import logging
 
 import torch
 
+import artist.nurbs.utils
+import artist.util.utils
+from artist.nurbs.surfaces import NURBSSurfaces
 from artist.scenario.configuration_classes import FacetConfig, SurfaceConfig
-from artist.util import config_dictionary, index_mapping, utils
+from artist.util import config_dictionary, index_mapping
 from artist.util.environment_setup import get_device
-from artist.util.nurbs import NURBSSurfaces
 
 log = logging.getLogger(__name__)
 """A logger for the surface generator."""
@@ -172,7 +174,7 @@ class SurfaceGenerator:
         control_points[:, :, :, :, index_mapping.u] = 0
 
         # Since NURBS are only defined between (0,1), we need to normalize the evaluation points and remove the boundary points.
-        evaluation_points[:, : index_mapping.u] = utils.normalize_points(
+        evaluation_points[:, : index_mapping.u] = artist.util.utils.normalize_points(
             evaluation_points[:, : index_mapping.u]
         )
         evaluation_points = evaluation_points.unsqueeze(0).unsqueeze(0)
@@ -325,11 +327,13 @@ class SurfaceGenerator:
             )
 
         # Convert to 4D format.
-        surface_points_with_facets = utils.convert_3d_points_to_4d_format(
+        surface_points_with_facets = artist.util.utils.convert_3d_points_to_4d_format(
             surface_points_with_facets, device=device
         )
-        surface_normals_with_facets = utils.convert_3d_directions_to_4d_format(
-            surface_normals_with_facets, device=device
+        surface_normals_with_facets = (
+            artist.util.utils.convert_3d_directions_to_4d_format(
+                surface_normals_with_facets, device=device
+            )
         )
 
         # Generate NURBS surface from multiple facets.
@@ -413,7 +417,7 @@ class SurfaceGenerator:
         log.info("Beginning generation of the ideal surface configuration.")
         facet_config_list = []
 
-        control_points = utils.create_planar_nurbs_control_points(
+        control_points = artist.nurbs.utils.create_planar_nurbs_control_points(
             number_of_control_points=self.number_of_control_points,
             canting=canting,
             device=device,
